requirements.txt
# ReconXploit v3.0 - Python Dependencies
# Product of Kernelpanic under infosbios.tech

# Core dependencies
asyncio>=3.4.3
aiohttp>=3.8.0
requests>=2.28.0
pyyaml>=6.0
jinja2>=3.1.0
click>=8.1.0
colorama>=0.4.6
python-dateutil>=2.8.2

# Web and HTTP
beautifulsoup4>=4.11.0
lxml>=4.9.0
urllib3>=1.26.0

# Data processing
pandas>=1.5.0
jsonschema>=4.17.0

# Networking and DNS
dnspython>=2.3.0

# Security and crypto
cryptography>=39.0.0

# Report generation
matplotlib>=3.6.0
pillow>=9.4.0

# API integrations
python-whois>=0.8.0

# Additional utilities
tqdm>=4.64.0
psutil>=5.9.0
tabulate>=0.9.0
